fix opcion eliminar lugar y documentos vehiculo

+++ database/migrations/2017_06_25_203506_create_pago_rel_rutas_table.php
@@ -23,7 +23,7 @@ class CreatePagoRelRutasTable extends Migration
             $table->foreign('ruta_id')->references('id')->on('rutas');
 
             $table->decimal('valor_final', 19, 2);
-            $table->integer('cantidad_viajes');
+            $table->decimal('cantidad_viajes', 3, 1);
             $table->integer('user_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
